Refactor: Retire /policy/check route, re-home diagnostic as library (#2500)
Unify the two policy entry points into one: /apply is now the sole policy
route. The read-only /policy/check route and its PolicyCheckRequest model are
removed from the controller.

The rich conflict-diagnostic engine is preserved and re-homed as an internal
library (not deleted, reversing handoff 07 step 4 per issue #2500):

  - Move the survey orchestrator check_policy_conflicts from the route-tied
    uc/policy_check/check.py into policy_rules_builder/diagnostic_survey.py,
    next to the per-entity engine; dissolve the orphaned uc/policy_check/ pkg.
  - diagnostic.py, diagnostic_models.py, prompts.py already lived in the
    library; drop the last textual tie to the removed route from the
    diagnostic_models docstring.

Tests mirror src: delete the route-level test, move + retarget the survey and
live-LLM tests to policy_rules_builder/, and move the /apply contradiction
regression guard (no symbol change). The live /apply proposer/auditor graph is
byte-for-byte unchanged.

Verification: unit suite 652 passed (no --ignore); llm suite 8 passed;
integration errors are pre-existing LLM auditor non-convergence in the live
/apply onboarding path, unrelated to this route retirement.
