+++ aiac/src/aiac/agent/policy_rules_builder/diagnostic_models.py
@@ -1,7 +1,7 @@
 """Structured models for the read-only policy **conflict-check** diagnostic (feature #154).
 
-This module defines ONLY the stable serialization shape shared by the diagnostic engine, the
-survey use-case, and the ``POST /policy/check`` route — no pipeline logic. The live ``/apply``
+This module defines ONLY the stable serialization shape shared by the diagnostic engine and the
+survey orchestrator — no pipeline logic. The live ``/apply``
 path and its models (``PolicyRule``, ``Contradiction``, ``AuditVerdict`` in ``graph.py`` /
 ``policy.model.models``) are deliberately untouched (decision D11): the typed conflict ``kind``
 lives here, produced by the diagnostic's ``explain`` node, rather than being bolted onto the
