--- aiac/src/aiac/agent/uc/policy_check/__init__.py
@@ -1,1 +1,0 @@
-"""Read-only Policy Conflict Check survey use-case (feature #154, task #158)."""
